+++ tests/settings.py
@@ -73,6 +73,7 @@ INSTALLED_APPS = (
     'wagtail.documents',
     'wagtail.admin',
     'wagtail.core',
+    'wagtail.contrib.redirects',
 
     'taggit',