+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.3.0 (04-10-2019)
+
+- Add `WAGTAIL_MARKETING_SEO_SCORE_ICONS` for overriding the current icon set.
+
+# 0.2.0 (12-08-2019)
+
+- Add the bulk redirect import on top of wagtail.contrib.redirects
+
 # 0.1.0 (15-04-2019)
 
 - No breaking changes, updated classifiers for a proper stable release